Clean up typings and code to make it more readable
This patch relaxes the typechecking and linting settings such that the
code can be made a bit more readable.

It removes the strict typechecking mode from tsconfig.json, and enforces
single quotes instead of double quites, as well as removal of curlies
around single-expression arrow functions.

The patch also allows for using type aliases instead of interfaces. We
don't extend and otherwise combine interfaces so using type aliases is
appropriate.