Implement patchless actions
The previous API for handling actions was a bit too sealed. This patch
modifies the API such that a single function is used instead of the
action-handler mapping. The mapping (if any) is now a user concern, and
duckweed won't do anything to match the messag address to its handler
function. In fact, address-less patterns are now possible.

This change also affects the plugin and middleware API, and the
functions previously used internally for patching are now exposed as the
`scoped` module to make working with the new API easier.

Documentation is updated to reflect the new API changes. Previously
missing middleware guide has now been added. We've also used the
opportunity to fix typos and technical omissions in the docs.

Closes #5
Closes #4
Closes #3